+++ forui/lib/src/widgets/line_calendar/line_calendar_item.dart
@@ -6,19 +6,8 @@ import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
 
-part 'line_calendar_item.style.dart';
-
 /// The state of a line calendar item used to build a line calendar item.
-typedef FLineCalendarItemData =
-    ({
-      FLineCalendarStyle style,
-      FLineCalendarItemStyle itemStyle,
-      DateTime date,
-      bool today,
-      bool selected,
-      bool hovered,
-      bool focused,
-    });
+typedef FLineCalendarItemData = ({FLineCalendarStyle style, Set<WidgetState> states, DateTime date, bool today});
 
 @internal
 class Item extends StatelessWidget {
@@ -41,40 +30,22 @@ class Item extends StatelessWidget {
   Widget build(BuildContext context) => ValueListenableBuilder(
     valueListenable: controller,
     builder:
-        (context, selected, _) => FTappable.animated(
+        (context, selectedDate, _) => FTappable.animated(
           focusNode: FocusNode(debugLabel: 'Date: $date'),
           semanticLabel: (FLocalizations.of(context) ?? FDefaultLocalizations()).fullDate(date),
           onPress: () => controller.select(date),
-          builder: (context, state, _) {
-            final itemStyle = switch ((selected == date, state.hovered)) {
-              (true, true) => style.selectedHoveredItemStyle,
-              (true, false) => style.selectedItemStyle,
-              (false, true) => style.unselectedHoveredItemStyle,
-              (false, false) => style.unselectedItemStyle,
-            };
+          builder: (context, states, _) {
+            final selected = selectedDate == date;
+            if (selected) {
+              states = {...states, WidgetState.selected};
+            }
 
             return builder(
               context,
-              (
-                style: style,
-                itemStyle: itemStyle,
-                date: date,
-                today: today,
-                selected: selected == date,
-                hovered: state.hovered,
-                focused: state.focused,
-              ),
+              (style: style, states: states, date: date, today: today),
               Stack(
                 children: [
-                  Positioned.fill(
-                    child: ItemContent(
-                      style: style,
-                      itemStyle: itemStyle,
-                      date: date,
-                      hovered: state.hovered,
-                      focused: state.focused,
-                    ),
-                  ),
+                  Positioned.fill(child: ItemContent(style: style, states: states, date: date)),
                   if (today)
                     Positioned(
                       top: 6,
@@ -82,7 +53,10 @@ class Item extends StatelessWidget {
                       child: Container(
                         height: 4,
                         width: 4,
-                        decoration: BoxDecoration(color: itemStyle.todayIndicatorColor, shape: BoxShape.circle),
+                        decoration: BoxDecoration(
+                          color: style.todayIndicatorColor.resolve(states),
+                          shape: BoxShape.circle,
+                        ),
                       ),
                     ),
                 ],
@@ -107,38 +81,29 @@ class Item extends StatelessWidget {
 @internal
 class ItemContent extends StatelessWidget {
   final FLineCalendarStyle style;
-  final FLineCalendarItemStyle itemStyle;
+  final Set<WidgetState> states;
   final DateTime date;
-  final bool hovered;
-  final bool focused;
 
-  const ItemContent({
-    required this.style,
-    required this.itemStyle,
-    required this.date,
-    required this.hovered,
-    required this.focused,
-    super.key,
-  });
+  const ItemContent({required this.style, required this.states, required this.date, super.key});
 
   @override
   Widget build(BuildContext context) {
     final localizations = FLocalizations.of(context) ?? FDefaultLocalizations();
     return DecoratedBox(
-      decoration: focused ? itemStyle.focusedDecoration : itemStyle.decoration,
+      decoration: style.decoration.resolve(states),
       child: Padding(
-        padding: EdgeInsets.symmetric(vertical: style.itemContentEdgeSpacing),
+        padding: EdgeInsets.symmetric(vertical: style.contentEdgeSpacing),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           mainAxisSize: MainAxisSize.min,
-          spacing: style.itemContentSpacing,
+          spacing: style.contentSpacing,
           children: [
             DefaultTextStyle.merge(
               textHeightBehavior: const TextHeightBehavior(
                 applyHeightToFirstAscent: false,
                 applyHeightToLastDescent: false,
               ),
-              style: itemStyle.dateTextStyle,
+              style: style.dateTextStyle.resolve(states),
               child: Text(localizations.day(date)),
             ),
             DefaultTextStyle.merge(
@@ -146,7 +111,7 @@ class ItemContent extends StatelessWidget {
                 applyHeightToFirstAscent: false,
                 applyHeightToLastDescent: false,
               ),
-              style: itemStyle.weekdayTextStyle,
+              style: style.weekdayTextStyle.resolve(states),
               child: Text(localizations.shortWeekDays[date.weekday % 7]),
             ),
           ],
@@ -160,41 +125,7 @@ class ItemContent extends StatelessWidget {
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty('style', style))
-      ..add(DiagnosticsProperty('itemStyle', itemStyle))
-      ..add(DiagnosticsProperty('date', date))
-      ..add(FlagProperty('hovered', value: hovered, ifTrue: 'hovered'))
-      ..add(FlagProperty('focused', value: focused, ifTrue: 'focused'));
+      ..add(IterableProperty('states', states))
+      ..add(DiagnosticsProperty('date', date));
   }
-}
-
-/// A line calendar item's state style.
-final class FLineCalendarItemStyle with Diagnosticable, _$FLineCalendarItemStyleFunctions {
-  /// The decoration.
-  @override
-  final BoxDecoration decoration;
-
-  /// The focused decoration.
-  @override
-  final BoxDecoration focusedDecoration;
-
-  /// The color of the today indicator.
-  @override
-  final Color todayIndicatorColor;
-
-  /// The text style for the date.
-  @override
-  final TextStyle dateTextStyle;
-
-  /// The text style for the day of the week.
-  @override
-  final TextStyle weekdayTextStyle;
-
-  /// Creates a [FLineCalendarItemStyle].
-  const FLineCalendarItemStyle({
-    required this.decoration,
-    required this.focusedDecoration,
-    required this.todayIndicatorColor,
-    required this.dateTextStyle,
-    required this.weekdayTextStyle,
-  });
 }